requirements.txt
Kivy==2.3.1
Pillow>=9.0.0
pywin32>=300
